Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-dotenv
pytest
pytest-asyncio
requests
jose
passlib
python-multipart